requirements.txt
django~=5.2.5
djangorestframework~=3.16.1
flake8
flake8-quotes
flake8-variables-names
django-debug-toolbar==6.0.0
pillow
djangorestframework-simplejwt
drf-spectacular~=0.28.0
django-filter~=25.1
django-environ~=0.12.0
python-dotenv~=1.1.1
coverage
requests~=2.32.5
dotenv
django-schema-viewer